requirements.txt
torch
torchvision
torchaudio
numpy
datasets
transformers[torch]
evaluate
tokenizers
msvc-runtime
matplotlib
aiogram
python-decouple